requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.24
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
email-validator>=2.0.0
